refactor: 型とアトムを整理
